requirements.txt
python-telegram-bot==22.6
yt-dlp
python-dotenv==1.0.1
aiofiles==23.2.1
